requirements.txt
# python utilities
requests
python-dotenv

# finance service providers sdk-s
alpaca-py
yfinance
